+++ HISTORY.rst
@@ -73,6 +73,7 @@ History
      * Fix Pool and Professional association relationships and update tests (rudaporto).
      * Update initial database migration script with all model changes (rudaporto).
      * LEICA-128: Refactor Job classes names following the changes as Assignment or Order (rudaporto).
+     * LEICA-132: Add new fields to Pool and fix Professionls in Pool import. Add pool attribute to Assignment list and filter (rudaporto).
 
 
 1.1.0 (2016-10-04)